Implement Create Payment Intent Endpoint #5: -Refactored packages and classes to align with Clean Arch style.
