--- src/test/java/devcunha/springboot/stripe/service/PaymentServiceTest.java
@@ -1,5 +1,0 @@
-package devcunha.springboot.stripe.service;
-
-public class PaymentServiceTest {
-
-}
