+++ src/test/java/devcunha/springboot/stripe/usecases/createpayment/CreatePaymentInteractorTest.java
@@ -1,0 +1,5 @@
+package devcunha.springboot.stripe.usecases.createpayment;
+
+public class CreatePaymentInteractorTest {
+
+}
